Add libpulseaudio and pulseaudio to dependencies; update wrapProgram in flake.nix
This commit enhances the flake.nix configuration by adding libpulseaudio and pulseaudio to the build inputs. It also refines the wrapProgram command to include additional environment variables for better integration with the system's D-Bus and GObject introspection. Additionally, a NixOS module is introduced to enable the power-profiles-daemon service, streamlining its integration into the system.